Split out unicode_ljust and unicode_rjust to break import cycle
Previously, there was an import cycle between tui.menu_item and
lib.output.
